@@ -60,8 +60,8 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-prettier": "^3.0.0",
-    "eslint-plugin-react": "^7.11.0",
-    "eslint-plugin-react-hooks": "^1.3.0",
+    "eslint-plugin-react": "^7.12.4",
+    "eslint-plugin-react-hooks": "^1.6.0",
     "eslint-plugin-you-dont-need-momentjs": "^1.4.0",
     "fast-glob": "^2.2.6",
     "file-loader": "^3.0.1",
@@ -80,7 +80,7 @@
     "ora": "^3.0.0",
     "prettier": "^1.16.1",
     "prettier-stylelint": "^0.4.2",
-    "react-test-renderer": "^16.7.0",
+    "react-test-renderer": "^16.8.6",
     "style-loader": "^0.23.0",
     "stylelint": "^9.9.0",
     "stylelint-config-recommended": "^2.1.0",
@@ -108,8 +108,8 @@
     "offline-plugin": "^5.0.5",
     "polished": "^3.0.0",
     "prop-types": "^15.6.1",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-fastclick": "^3.0.2-alpha.1",
     "react-helmet": "^5.0.3",
     "react-hot-loader": "^4.3.11",
